refactor(storage): break meta-storage and transformer config cycles
Extract leaf-level type files to eliminate all 4 file cycles flagged by
madge in @visulima/storage:

- storage/meta-storage-options.ts: MetaStorageOptions + LocalMetaStorageOptions
- transformer/transformer-config.ts: BaseTransformerConfig

Runtime files (meta-storage.ts, local-meta-storage.ts, base-transformer.ts)
import from the leaf types instead of from the main types barrels, while
types barrels re-export for backward compat.
